refactor(rules): split helpers.go into topic-specific files
Breaks the 323-line grab bag into four smaller files:
- helpers_path.go: path/module/exclude resolution
- helpers_sublayer.go: port/contract sublayer identification
- helpers_ast.go: AST traversal helpers
- run_all.go: deduplicateMetaViolations moves next to its only caller

No behavior change; all tests and lint green.
